monitors/pyproject.toml
[project]
name = "monitor-manager"
version = "0.1.0"
description = "macOS window position manager for multiple monitors"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyobjc-core>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-ApplicationServices>=10.0",
]

[project.scripts]
monitor-manager = "monitor_manager.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
